requirements/base.txt

celery==4.0.2

Django==1.11.1

django-filter==1.0.2
django-rest-swagger==2.1.2
djangorestframework==3.6.3

whitenoise==3.3.0
dj-database-url==0.4.2
dj-email-url==0.0.10

psycopg2==2.7.1



django-redis==4.8.0
redis==2.10.5


